requirements.txt
pytest>=7.1.2
allure-pytest>=2.9.45
requests>=2.31.0
python-dotenv~=1.0.0
requests>=2.31.0
jsonschema>=4.20.0